todolist v2 is fully working, create new list and use dynamic routing…

+++ projects/todolist-v2/app.js
@@ -3,6 +3,7 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const date = require(`${__dirname}/date.js`);
 const mongoose = require('mongoose');
+const _ = require("lodash");
 
 const day = date();
 
@@ -119,13 +120,32 @@ app.post("/", (req,res) => {
 app.post("/delete", (req, res) => {
 
     const checkedItemId = req.body.checkbox;
+    const listName = req.body.listName;
 
-    Item.findByIdAndRemove(checkedItemId, (err) => {
-        if(err) console.log(err);
-        console.log("Successful Deletion");
-    })
+    if(listName === "Today"){
 
-    res.redirect("/")
+        Item.findByIdAndRemove(checkedItemId, (err) => {
+            if(err) console.log(err);
+            console.log("Successful Deletion");
+        })
+    
+        res.redirect("/")
+
+    } else {
+
+        List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}}, (err, foundList) => {
+
+            if(!err){
+                res.redirect(`/${listName}`);
+            } else {
+                console.log(err);
+            }
+
+        })
+
+    }
+
+
 
 })
 
@@ -134,7 +154,7 @@ app.post("/delete", (req, res) => {
  */
 app.get("/:customListName", (req, res) => {
 
-    const customListName = req.params.customListName;
+    const customListName = _.capitalize(req.params.customListName);
 
     List.findOne({name: customListName}, (err, foundList) => {
         if(!err) {
